Rename test/simplecov.rb to test/simplecov_init.rb so that bundle exec rake works with bundler version 1.12.5
